move remaining function to IO and remove Settings
