+++ mrtarget/common/IO.py
@@ -63,3 +63,20 @@ def make_iter_lines(iterable_of_filenames, first_n=0):
 
     return more_itertools.take(first_n, it_lines) \
         if first_n > 0 else it_lines
+
+
+def file_or_resource(fname):
+    '''get filename and check if in getcwd then get from
+    the package resources folder
+    '''
+    filename = os.path.expanduser(fname)
+
+    resource_package = mrtarget.__name__
+    resource_path = '/'.join(('resources', filename))
+
+    if filename is not None:
+        abs_filename = os.path.join(os.path.abspath(os.getcwd()), filename) \
+                       if not os.path.isabs(filename) else filename
+
+        return abs_filename if os.path.isfile(abs_filename) \
+            else res.resource_filename(resource_package, resource_path)
